Closes #99 + detection of user departure and destination drop
The user can only successfully move a layer once. Then, issue #97
causes bugs.
Also, for #99, JS detects `departureLayerIndex` as a string and appends
`i` without a forced parse. When first operated on with a subtraction,
this issue disappears as it recognizes it properly as an integer. See
line 289’s LHS vs RHS.