Create LICENSE
The license was already created, but somehow vanished. I suspect it might have been due to not pulling from master before merging at some point? I am unsure. But in any case, I cannot find the original license which was supposed to have been done June 28, 2017. In the commit history, it does not appear.

This should be an exact copy of what was done originally.